Throttle.cc: try to avoid a negative wait time and thread starvation
Start + delay could be less than now(), and thus could wait_until be
called with the timepoint that is in the past.
It is uncertain what different OSes guarantee in this case.
On FreeBSD is seems that it leads to no (or very little) possibility to switch thread.
And thus throttle.put() with starve, which would lead to no progress.

Signed-off-by: Willem Jan Withagen <wjw@digiware.nl>